add test to chart and headline components
